Move rsdp_search_bios etc to rsdp_search and a few changes
